Реализация UserController и UserService

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -10,8 +10,8 @@ import java.util.Map;
 @Component
 public class InMemoryUserStorage implements UserStorage {
 
-    private final Map<Integer, User> users = new HashMap<>();
-    private static int index = 0;
+    private final Map<Long, User> users = new HashMap<>();
+    private static long index = 0;
 
     @Override
     public User add(User user) {
@@ -44,7 +44,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public Map<Integer, User> getAllUser() {
+    public Map<Long, User> getAllUser() {
         return users;
     }
 }
